+++ teamcity-rest-client-impl/src/main/kotlin/org/jetbrains/teamcity/rest/coroutines/rest.kt
@@ -659,7 +659,6 @@ internal class ArtifactDependenciesBean {
 }
 
 internal class SnapshotDependencyBean: IdBean() {
-    var name: String? = null
     var disabled: Boolean? = false
     var inherited: Boolean? = false
     var properties: ParametersBean? = ParametersBean()