+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to teamcity-rest-client library will be documented in this file.
 
 ## [Unreleased]
-- Added ability to fetch builds' snapshot dependencies
+- Added ability to fetch build configurations' snapshot dependencies
 - Added property `type` to `BuildConfiguration`
 - Added a way to specify precise logging level for HTTP requests